Support straight tails--and multiple tail types in general
